[UI] Add consumed / produced values on progress bars [Notifications] Trigger notifications when out of beer (no trigger if time < 24h)

+++ Assets/Scripts/UI/ProperUIScrpits/Header/AllDataDisplayer.cs
@@ -68,7 +68,7 @@
 
 		private void UpdateBeerDisplay()
 		{
-			string displayText = UIHelper.FormatIntegerString(_playerProfile.CurrentFortress.Beer);
+			string displayText = UIHelper.FormatIntegerString((int)_playerProfile.CurrentFortress.Beer);
 			_beerDisplay.text = displayText;
 		}
